Verify file does not exist before creating symlink

+++ dracut/modules.d/90kiwi-live/kiwi-generator.sh
@@ -1,4 +1,4 @@
-#!/bin/bash
+#!/bin/bash -x
 
 type getarg >/dev/null 2>&1 || . /lib/dracut-lib.sh
 
@@ -36,9 +36,11 @@ ROOTFLAGS="$(getarg rootflags)"
     _dev=LiveOS_rootfs
 } > "$GENERATOR_DIR"/sysroot.mount
 
-mkdir -p "$GENERATOR_DIR"/initrd-root-fs.target.requires
-ln -s "$GENERATOR_DIR"/sysroot.mount \
-    "$GENERATOR_DIR"/initrd-root-fs.target.requires/sysroot.mount
+if [ ! -e "$GENERATOR_DIR/initrd-root-fs.target.requires/sysroot.mount" ]; then
+    mkdir -p "$GENERATOR_DIR"/initrd-root-fs.target.requires
+    ln -s "$GENERATOR_DIR"/sysroot.mount \
+        "$GENERATOR_DIR"/initrd-root-fs.target.requires/sysroot.mount
+fi
 
 mkdir -p "$GENERATOR_DIR/$_dev.device.d"
 {
